Fix solution visualization sudden "deflation"
Globally refine twice instead of just once in Step6 constructor, which did solve the issue of the solution visualization appearing to suddenly deflate after the 6th cycle. No actual sudden decrease in solution values was occurring, but the scaling of the z-axis drastically changed after the 6th cycle for unknown reasons with just one global refinement in the constructor.