assemble_system() independent of cycle
The first element of each solutionfunction_vector is now the zero function. This way, we no longer have work around the fact that the solutionfunction_vectors were empty at the beginning of the very first cycle. Therefore, we now need the last element of the solutionfunction_vector corresponding to the relevant_subdomain whether we are in the first cycle or not, making assemble_system() independent of the cycle number.