Incomplete correction for singularities
Initially, I was imposing boundary conditions that were not guaranteed to be continuous. To overcome this, in the 2x2 orientation of our four subdomains, the solutions of subdomains diagonal to each other need to be considered. I needed to split edges of each subdomain into two segments having different boundary_ids in order to be able to impose different solutions on each of these segments of the same edge. Currently, errors are present and this does not yet run properly.