Remove some std::fabs() in setting boundary_ids
When setting boundary_ids in the constructor, the second condition for each if statement wrongly used std::fabs(). This has been removed because for the second condition, I am not looking at the absolute difference between two quantities, but instead am looking at a range of values in the coordinate direction not included in the first if statement condition.