Beginning of implentation of Additive Schwarz
Additive Schwarz implementation not yet complete: BC from regions of overlap need to be some average of the solutions from the overlapping subdomains.
Names of output files need to be changed to indicate whether additive or multiplicative was used. File names that are commented out led to errors when attempting to visualize in visIt.