Incomplete correction to Additive Schwarz
Before BC entirely retrieved from neighboring subdomain solution i.e. the only difference between my previous implementation of Additive and Multiplicative Schwarz was that the solution imposed in MS was the most recently computed one while for AS it was the solution from the previous cycle. I have begun changes to instead use a weighted average of all solutions from subdomains in the overlapping regions.